Delete old API objects
The `ProjectConfig` class is no longer used, as is the `ConfigSchema`
class
