--- src/maison/schema.py
@@ -1,6 +1,0 @@
-"""Module to define the `ConfigSchema` class."""
-from pydantic import BaseModel
-
-
-class ConfigSchema(BaseModel):
-    """A class for creating schemas based on the `pydantic` `BaseModel`."""
